requirements/base.txt
psycopg2-binary
django-rest-framework

